requirements.txt
mitmproxy
discord.py
aiohttp
requests
pycryptodome
protobuf-decoder

